feat(compass-indexes): add index build progress COMPASS-9495

+++ packages/compass-indexes/src/components/regular-indexes-table/regular-index-actions.tsx
@@ -1,11 +1,30 @@
 import semver from 'semver';
 import React, { useCallback, useMemo } from 'react';
 import type { GroupedItemAction } from '@mongodb-js/compass-components';
-import { css, ItemActionGroup } from '@mongodb-js/compass-components';
+import {
+  css,
+  ItemActionGroup,
+  SpinLoader,
+  spacing,
+  Body,
+} from '@mongodb-js/compass-components';
 
 const styles = css({
   // Align actions with the end of the table
   justifyContent: 'flex-end',
+});
+
+const combinedContainerStyles = css({
+  display: 'flex',
+  alignItems: 'center',
+  gap: spacing[200],
+  minWidth: spacing[800],
+  justifyContent: 'flex-end',
+});
+
+const progressTextStyles = css({
+  fontSize: '12px',
+  fontWeight: 'normal',
 });
 
 type Index = {
@@ -13,6 +32,8 @@ type Index = {
   extra?: {
     hidden?: boolean;
   };
+  status?: 'inprogress' | 'ready' | 'failed';
+  progressPercentage?: number;
 };
 
 type IndexActionsProps = {
@@ -44,30 +65,42 @@ const IndexActions: React.FunctionComponent<IndexActionsProps> = ({
 }) => {
   const indexActions: GroupedItemAction<IndexAction>[] = useMemo(() => {
     const actions: GroupedItemAction<IndexAction>[] = [];
+    const progressPercentage = index.progressPercentage ?? 0;
 
-    if (serverSupportsHideIndex(serverVersion)) {
-      actions.push(
-        index.extra?.hidden
-          ? {
-              action: 'unhide',
-              label: `Unhide Index ${index.name}`,
-              tooltip: `Unhide Index`,
-              icon: 'Visibility',
-            }
-          : {
-              action: 'hide',
-              label: `Hide Index ${index.name}`,
-              tooltip: `Hide Index`,
-              icon: 'VisibilityOff',
-            }
-      );
+    if (progressPercentage < 100 && progressPercentage > 0) {
+      // partially built
+      actions.push({
+        action: 'delete',
+        label: `Cancel Index ${index.name}`,
+        icon: 'XWithCircle',
+        variant: 'destructive',
+      });
+    } else {
+      // completed
+      if (serverSupportsHideIndex(serverVersion)) {
+        actions.push(
+          index.extra?.hidden
+            ? {
+                action: 'unhide',
+                label: `Unhide Index ${index.name}`,
+                tooltip: `Unhide Index`,
+                icon: 'Visibility',
+              }
+            : {
+                action: 'hide',
+                label: `Hide Index ${index.name}`,
+                tooltip: `Hide Index`,
+                icon: 'VisibilityOff',
+              }
+        );
+      }
+
+      actions.push({
+        action: 'delete',
+        label: `Drop Index ${index.name}`,
+        icon: 'Trash',
+      });
     }
-
-    actions.push({
-      action: 'delete',
-      label: `Drop Index ${index.name}`,
-      icon: 'Trash',
-    });
 
     return actions;
   }, [index, serverVersion]);
@@ -84,6 +117,30 @@ const IndexActions: React.FunctionComponent<IndexActionsProps> = ({
     },
     [onDeleteIndexClick, onHideIndexClick, onUnhideIndexClick, index]
   );
+
+  const progressPercentage = index.progressPercentage ?? 0;
+  if (
+    index.status !== 'failed' &&
+    progressPercentage < 100 &&
+    progressPercentage > 0
+  ) {
+    const progressText = `${Math.round(progressPercentage)}%`;
+
+    return (
+      <div
+        className={combinedContainerStyles}
+        data-testid="index-building-spinner"
+      >
+        <Body className={progressTextStyles}>Building... {progressText}</Body>
+        <SpinLoader size={16} title="Index build in progress" />
+        <ItemActionGroup<IndexAction>
+          data-testid="index-actions"
+          actions={indexActions}
+          onAction={onAction}
+        />
+      </div>
+    );
+  }
 
   return (
     <ItemActionGroup<IndexAction>
